feedback: move Htmx debug tools into its own mixin
